Updated Presentation API shim to latest version of the spec
The shim now exposes the most recent interfaces, implementing in particular
the division between PresentationRequest and PresentationReceiver for the
controlling browsing context and the receiving browsing context.

Actually, the shim implements changes that are not yet in the spec:
- PresentationSession is rather called PresentationConnection in the shim
- getSession and getSessions were renamed getConnection and getConnections
accordingly.

This might need to be revisited depending on what the specs says in fine.

The shim has also been re-written to follow the algorithms defined in the spec
more closely, when possible. It prompts the user to select a display (although
it's more a choice between presentation mechanisms than a choice of available
displays for the time being).

New presentation mechanisms need to implement a "getAvailableDisplays" and a
"monitorIncomingContexts" methods. Displays are represented internally as
so-called remote "BrowsingContext" objects. The name is a bit mis-leading
since these objects do not really represent browsing contexts in practice. They
don't really represent displays either.

There are two types of browsing contexts: those that can be navigated called
ReceivingBrowsingContext, and those that can't. Presentation mechanisms will
typically implement a version of each for their own mechanism.

The shim exposes itself with a w3c prefix to avoid collision with recent dev
versions of browsers that ship with a native implementation of the Presentation
API.