New DataChannel, RemoteController, Display internal interfaces
These classes replace the former BrowsingContext and ReceivingBrowsingContext
interfaces which were badly named. RemoteController and Display inherit from
DataChannel (actually RemoteController is nothing more than a DataChannel in
practice).

The PresentationConnection constructor now takes a DataChannel as parameter.
Depending on where this connection gets created, this will either be a
RemoteController or a Display.

The code also now uses the new "closed" state instead of "disconnected".